FIXED IP-286 - Discount input fields are not disabled on read-only

+++ application/modules/invoices/views/partial_item_table.php
@@ -160,10 +160,12 @@
                     </div>
                 </td>
                 <td class="td-icon text-right td-vert-middle">
-                    <a href="<?php echo site_url('invoices/delete_item/' . $invoice->invoice_id . '/' . $item->item_id); ?>"
-                       title="<?php echo lang('delete'); ?>">
-                        <i class="fa fa-trash-o text-danger"></i>
-                    </a>
+                    <?php if ($invoice->is_read_only != 1): ?>
+                        <a href="<?php echo site_url('invoices/delete_item/' . $invoice->invoice_id . '/' . $item->item_id); ?>"
+                           title="<?php echo lang('delete'); ?>">
+                            <i class="fa fa-trash-o text-danger"></i>
+                        </a>
+                    <?php endif; ?>
                 </td>
             </tr>
             <tr>
